Add files via upload
This is a basic template we are going to be using for the UI for the website. This has some reactive theme that  will show you what page you are in by matching the background color to the function. 


feel free to add buttons and further implement other design elements and functions